refactor: 위경도 decimal에서 float으로 변경

+++ post/schemas.py
@@ -1,13 +1,12 @@
 from datetime import datetime
-from decimal import Decimal
 
 from ninja import Schema
 
 
 class PostCreateSchema(Schema):
     text: str
-    latitude: Decimal
-    longitude: Decimal
+    latitude: float
+    longitude: float
 
 
 class PostSchema(Schema):
@@ -19,8 +18,8 @@ class PostSchema(Schema):
     created_at: datetime
     updated_at: datetime
     like_count: int
-    latitude: Decimal
-    longitude: Decimal
+    latitude: float
+    longitude: float
     is_liked: bool
 
     @staticmethod
@@ -35,4 +34,4 @@ class PostSchema(Schema):
 
 
 class PostWithDistanceSchema(PostSchema):
-    distance: Decimal
+    distance: float
